[fix]gridsizeをdoubleへ

+++ ImageRecognition/mygraphicsview.h
@@ -32,7 +32,7 @@ private:
     cv::Mat image;
     polygon_t polygon;
     std::vector<QPointF> points;
-    int grid_size;
+    double grid_size;
     int left_right_margin,top_buttom_margin;
     std::pair<double,double> minXY;
     bool selecting;
